% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels.all}
\alias{vowels.all}
\title{This dataset is most useful for normalization purposes, since it
contains data on both stressed and post-tonic vowels in a single
format}
\format{\preformatted{'data.frame':	5534 obs. of  19 variables:
 $ speaker    : Factor w/ 8 levels "Sp1","Sp2","Sp3",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ v          : Factor w/ 6 levels "i","\\\\sw","e",..: 4 1 4 3 3 2 2 1 4 5 ...
 $ v.f1       : num  422 517 709 943 813 ...
 $ v.f2       : num  655 2314 997 2345 1549 ...
 $ v.dur      : num  0.103 0.0279 0.0355 0.0421 0.0576 ...
 $ vh.dur     : num  0.103 0.0549 0.101 0.0692 0.1043 ...
 $ context    : Factor w/ 5 levels "short","cluster",..: 3 1 1 1 1 4 2 3 3 1 ...
 $ pos        : chr  "v1" "v1" "v1" "v1" ...
 $ ctx        : Factor w/ 3 levels "short","long",..: 2 1 1 1 1 1 1 2 2 1 ...
 $ v.f1.mean  : num  578 578 578 578 578 ...
 $ v.f2.mean  : num  1931 1931 1931 1931 1931 ...
 $ v.f1.sd    : num  220 220 220 220 220 ...
 $ v.f2.sd    : num  789 789 789 789 789 ...
 $ v.dur.mean : num  0.13 0.13 0.13 0.13 0.13 ...
 $ v.dur.sd   : num  0.0596 0.0596 0.0596 0.0596 0.0596 ...
 $ vh.dur.mean: num  0.136 0.136 0.136 0.136 0.136 ...
 $ vh.dur.sd  : num  0.0544 0.0544 0.0544 0.0544 0.0544 ...
 $ v.dur.norm : num  -0.445 -1.705 -1.577 -1.466 -1.206 ...
 $ vh.dur.norm: num  -0.599 -1.483 -0.636 -1.221 -0.575 ...
}}
\usage{
vowels.all
}
\description{
This dataset is most useful for normalization purposes, since it
contains data on both stressed and post-tonic vowels in a single
format
}
\keyword{datasets}

